[FLINK-38222] Bump HBase version from 2.2 to 2.6
