+++ .github/workflows/push_pr.yml
@@ -28,8 +28,8 @@ jobs:
   compile_and_test:
     strategy:
       matrix:
-        flink: [ 1.20.0 ]
-        jdk: [ '8, 11' ]
+        flink: [ 1.20.2 ]
+        jdk: [ '8, 11, 17' ]
     uses: apache/flink-connector-shared-utils/.github/workflows/ci.yml@ci_utils
     with:
       flink_version: ${{ matrix.flink }}
