+++ docs/content/docs/connectors/table/hbase.md
@@ -58,7 +58,7 @@ CREATE TABLE hTable (
  family3 ROW<q4 DOUBLE, q5 BOOLEAN, q6 STRING>,
  PRIMARY KEY (rowkey) NOT ENFORCED
 ) WITH (
- 'connector' = 'hbase-2.2',
+ 'connector' = 'hbase-2.6',
  'table-name' = 'mytable',
  'zookeeper.quorum' = 'localhost:2181'
 );
@@ -133,7 +133,7 @@ Connector Options
       <td>String</td>
       <td>Specify what connector to use, valid values are:
         <ul>
-            <li><code>hbase-2.2</code>: connect to HBase 2.2.x cluster</li>
+            <li><code>hbase-2.6</code>: connect to HBase 2.6.x cluster</li>
         </ul>
       </td>
     </tr>
@@ -225,7 +225,7 @@ Connector Options
       <td>no</td>
       <td style="word-wrap: break-word;">false</td>
       <td>Boolean</td>
-      <td>Whether async lookup are enabled. If true, the lookup will be async. Note, async only supports hbase-2.2 connector.</td>
+      <td>Whether async lookup are enabled. If true, the lookup will be async. Note, async only supports hbase-2.6 connector.</td>
     </tr>
     <tr>
       <td><h5>lookup.cache</h5></td>
