+++ docs/data/hbase.yml
@@ -17,7 +17,7 @@
 ################################################################################
 
 version: 4.0.0
-flink_compatibility: ["1.18", "1.19"]
+flink_compatibility: ["1.20"]
 variants:
-  - maven: flink-connector-hbase-2.2
-    sql_url: https://repo.maven.apache.org/maven2/org/apache/flink/flink-sql-connector-hbase-2.2/$full_version/flink-sql-connector-hbase-2.2-$full_version.jar
+  - maven: flink-connector-hbase-2.6
+    sql_url: https://repo.maven.apache.org/maven2/org/apache/flink/flink-sql-connector-hbase-2.6/$full_version/flink-sql-connector-hbase-2.6-$full_version.jar
