+++ flink-connector-hbase-2.6/src/test/java/org/apache/flink/architecture/TestCodeArchitectureTest.java
