+++ flink-connector-hbase-2.6/src/test/java/org/apache/flink/connector/hbase2/util/HBaseTestBase.java
