+++ flink-connector-hbase-2.6/src/test/java/org/slf4j/impl/Log4jLoggerAdapter.java
