+++ flink-connector-hbase-2.6/src/test/resources/archunit.properties
