+++ flink-connector-hbase-2.6/src/test/resources/log4j2-test.properties
