+++ flink-connector-hbase-base/src/test/resources/log4j.properties
@@ -16,31 +16,10 @@
 # limitations under the License.
 ################################################################################
 
-# We need to specify repo related information here since Apache INFRA doesn't differentiate
-# between several workflows with the same names while preparing a report for GHA usage
-# https://infra-reports.apache.org/#ghactions
-name: Weekly Flink Connector HBase CI
-on:
-  schedule:
-    - cron: "0 0 * * 0"
-  workflow_dispatch:
-jobs:
-  compile_and_test:
-    if: github.repository_owner == 'apache'
-    strategy:
-      matrix:
-        flink_branches: [{
-          flink: 1.20.2,
-          branch: main
-        }, {
-          flink: 1.19.3,
-          branch: v4.0
-        }]
-    uses: apache/flink-connector-shared-utils/.github/workflows/ci.yml@ci_utils
-    with:
-      flink_version: ${{ matrix.flink_branches.flink }}
-      connector_branch: ${{ matrix.flink_branches.branch }}
-      jdk_version: ${{ matrix.flink_branches.jdk || '8, 11, 17' }}
-      run_dependency_convergence: false
-      timeout_global: 120
-      timeout_test: 80
+# This is used only by Hadoop in the local tests, and we want to disable that to not flood the logs.
+log4j.rootLogger=ERROR, stdout
+
+log4j.appender.stdout=org.apache.log4j.ConsoleAppender
+log4j.appender.stdout.Target=System.out
+log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
+log4j.appender.stdout.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} %-5p %c{1}:%L - %m%n
