--- flink-sql-connector-hbase-2.2/src/main/resources/META-INF/NOTICE
@@ -1,94 +1,0 @@
-flink-sql-connector-hbase-2.2
-Copyright 2014-2024 The Apache Software Foundation
-
-This product includes software developed at
-The Apache Software Foundation (http://www.apache.org/).
-
-This project bundles the following dependencies under the Apache Software License 2.0. (http://www.apache.org/licenses/LICENSE-2.0.txt)
-
- - commons-codec:commons-codec:1.15
- - io.netty:netty-all:4.1.70.Final
- - io.netty:netty-buffer:4.1.70.Final
- - io.netty:netty-codec:4.1.70.Final
- - io.netty:netty-codec-dns:4.1.70.Final
- - io.netty:netty-codec-haproxy:4.1.70.Final
- - io.netty:netty-codec-http:4.1.70.Final
- - io.netty:netty-codec-http2:4.1.70.Final
- - io.netty:netty-codec-memcache:4.1.70.Final
- - io.netty:netty-codec-mqtt:4.1.70.Final
- - io.netty:netty-codec-redis:4.1.70.Final
- - io.netty:netty-codec-smtp:4.1.70.Final
- - io.netty:netty-codec-socks:4.1.70.Final
- - io.netty:netty-codec-stomp:4.1.70.Final
- - io.netty:netty-codec-xml:4.1.70.Final
- - io.netty:netty-common:4.1.70.Final
- - io.netty:netty-handler:4.1.70.Final
- - io.netty:netty-handler-proxy:4.1.70.Final
- - io.netty:netty-resolver:4.1.70.Final
- - io.netty:netty-resolver-dns:4.1.70.Final
- - io.netty:netty-resolver-dns-classes-macos:4.1.70.Final
- - io.netty:netty-resolver-dns-native-macos:osx-aarch_64:4.1.70.Final
- - io.netty:netty-resolver-dns-native-macos:osx-x86_64:4.1.70.Final
- - io.netty:netty-transport:4.1.70.Final
- - io.netty:netty-transport-classes-epoll:4.1.70.Final
- - io.netty:netty-transport-classes-kqueue:4.1.70.Final
- - io.netty:netty-transport-native-unix-common:4.1.70.Final
- - io.netty:netty-transport-native-epoll:linux-aarch_64:4.1.70.Final
- - io.netty:netty-transport-native-epoll:linux-x86_64:4.1.70.Final
- - io.netty:netty-transport-native-kqueue:osx-aarch_64:4.1.70.Final
- - io.netty:netty-transport-native-kqueue:osx-x86_64:4.1.70.Final
- - io.netty:netty-transport-rxtx:4.1.70.Final
- - io.netty:netty-transport-sctp:4.1.70.Final
- - io.netty:netty-transport-udt:4.1.70.Final
- - io.dropwizard.metrics:metrics-core:3.2.6
- - org.apache.commons:commons-crypto:1.0.0
- - org.apache.commons:commons-lang3:3.3.2
- - org.apache.hbase:hbase-client:2.2.3
- - org.apache.hbase:hbase-common:2.2.3
- - org.apache.hbase:hbase-protocol:2.2.3
- - org.apache.hbase:hbase-protocol-shaded:2.2.3
- - org.apache.hbase.thirdparty:hbase-shaded-protobuf:2.2.1
- - org.apache.hbase.thirdparty:hbase-shaded-miscellaneous:2.2.1
- - org.apache.hbase.thirdparty:hbase-shaded-netty:2.2.1
- - org.apache.htrace:htrace-core4:4.2.0-incubating
- - org.apache.zookeeper:zookeeper:3.4.14
-
-This project bundles the following dependencies under the BSD license.
-See bundled license files for details.
-
- - com.google.protobuf:protobuf-java:2.5.0
-
-The bundled Apache HTrace org.apache.htrace:htrace-core4 dependency bundles the following dependencies under
-the Apache Software License 2.0 (http://www.apache.org/licenses/LICENSE-2.0.txt)
-
- - com.fasterxml.jackson.core:jackson-annotations:2.4.0
- - com.fasterxml.jackson.core:jackson-core:2.4.0
- - com.fasterxml.jackson.core:jackson-databind:2.4.0
- - commons-logging:commons-logging:1.1.1
-
-The bundled Apache HBase Patched & Relocated (Shaded) Protobuf org.apache.hbase.thirdparty:hbase-shaded-protobuf
-dependency bundles the following dependencies under the BSD license. See bundled license files for details.
-
- - com.google.protobuf:protobuf-java:3.7.1
-
-The bundled Apache HBase Relocated (Shaded) Third-party Miscellaneous Libs
-org.apache.hbase.thirdparty:hbase-shaded-miscellaneous dependency bundles the following dependencies under
-the Apache Software License 2.0 (http://www.apache.org/licenses/LICENSE-2.0.txt)
-
- - com.google.code.gson:gson:2.8.5
- - com.google.guava:failureaccess:1.0.1
- - com.google.guava:guava:27.1-jre
- - commons-cli:commons-cli:1.4
- - org.apache.commons:commons-collections4:4.3
-
-The bundled Apache HBase Relocated (Shaded) Third-party Miscellaneous Libs
-org.apache.hbase.thirdparty:hbase-shaded-miscellaneous dependency bundles the following dependencies under
-the BSD license. See bundled license files for details.
-
- - com.google.protobuf:protobuf-java-util:3.7.1
-
-The bundled Apache HBase Relocated (Shaded) Netty Libs org.apache.hbase.thirdparty:hbase-shaded-netty
-dependency bundles the following dependencies under
-the Apache Software License 2.0 (http://www.apache.org/licenses/LICENSE-2.0.txt)
-
- - io.netty:netty-all:4.1.34.Final
