+++ flink-sql-connector-hbase-2.6/src/main/resources/META-INF/licenses/LICENSE.protobuf
